move utilities to apps module
