+++ busy_beaver/blueprints/slack/decorators.py
@@ -1,11 +1,9 @@
 import functools
 import hashlib
 import hmac
-from typing import List
 
 from flask import request
 
-from .toolbox import make_slack_response
 from busy_beaver.config import SLACK_SIGNING_SECRET
 from busy_beaver.exceptions import UnverifiedWebhookRequest
 
@@ -34,6 +32,9 @@ def verify_slack_signature(signing_secret: str):
     return verification_decorator
 
 
+slack_verification_required = verify_slack_signature(SLACK_SIGNING_SECRET)
+
+
 def calculate_signature(signing_secret, timestamp, data):
     req = str.encode("v0:" + str(timestamp) + ":") + data
     return "v0=" + hmac.new(str.encode(signing_secret), req, hashlib.sha256).hexdigest()
@@ -41,26 +42,3 @@ def calculate_signature(signing_secret, timestamp, data):
 
 def signatures_unequal(request_hash, slack_signature):
     return not hmac.compare_digest(request_hash, slack_signature)
-
-
-def limit_to(workspace_ids: List[str]):
-    """This decorator limits functionality to specified workspace_ids"""
-
-    if not isinstance(workspace_ids, list):
-        raise ValueError
-
-    def limit_to_decorator(func):
-        @functools.wraps(func)
-        def _wrapper(*args, **kwargs):
-            team_id = kwargs["team_id"]
-            if team_id not in workspace_ids:
-                return make_slack_response(text="Command not supported at this time.")
-
-            return func(*args, **kwargs)
-
-        return _wrapper
-
-    return limit_to_decorator
-
-
-slack_verification_required = verify_slack_signature(SLACK_SIGNING_SECRET)
